--- validator/evaluation/diffusion_eval_data.json
@@ -1,1 +1,0 @@
-{"test_split_path": "/workspace/input_data", "base_model_repo": "SG161222/RealVisXL_V4.0", "base_model_filename": "RealVisXL_V4.0.safetensors", "lora_repos": {"diagonalge/animelora": ["test/kohyatest.safetensors", "kohyatest-000018.safetensors"]}}
