+++ validator/evaluation/utils.py
@@ -63,7 +63,6 @@ def model_is_a_finetune(original_repo: str, finetuned_model: AutoModelForCausalL
     )
     return architecture_same and (base_model_match or has_lora_modules)
 
-
 def get_default_dataset_config(dataset_name: str) -> str | None:
     try:
         logger.info(dataset_name)
@@ -77,13 +76,11 @@ def get_default_dataset_config(dataset_name: str) -> str | None:
     else:
         return None
 
-
 def base64_to_image(base64_string: str) -> Image.Image:
     image_data = base64.b64decode(base64_string)
     image_stream = BytesIO(image_data)
     image = Image.open(image_stream)
     return image
-
 
 def download_from_huggingface(repo_id: str, filename: str, local_dir: str) -> str:
     # Use a temp folder to ensure correct file placement
@@ -101,11 +98,18 @@ def download_from_huggingface(repo_id: str, filename: str, local_dir: str) -> st
     except Exception as e:
         logger.error(f"Error downloading file: {e}")
 
+def list_supported_images(dataset_path: str, extensions: tuple) -> list[str]:
+    return [
+        file_name for file_name in os.listdir(dataset_path)
+        if file_name.lower().endswith(extensions)
+    ]
 
-def calculate_l2_loss(test_image: Image.Image, generated_image: Image.Image) -> float:
-    test_image = np.array(test_image.convert("RGB")) / 255.0
-    generated_image = np.array(generated_image.convert("RGB")) / 255.0
-    if test_image.shape != generated_image.shape:
-        raise ValueError("Images must have the same dimensions to calculate L2 loss.")
-    l2_loss = np.mean((test_image - generated_image) ** 2)
-    return l2_loss
+def read_image_as_base64(image_path: str) -> str:
+    with open(image_path, "rb") as image_file:
+        return base64.b64encode(image_file.read()).decode("utf-8")
+
+def read_prompt_file(text_file_path: str) -> str:
+    if os.path.exists(text_file_path):
+        with open(text_file_path, "r", encoding="utf-8") as text_file:
+            return text_file.read()
+    return None
